perf: Optimize render performance
BEFORE:
On every value update, the complete form rerendered. This was partially
intended as the complete form api is available in a Field and may be
used to change the appearance of the field.
NOW:
As in 99% of the cases you don't need to update a field if another field
changes, a field now only renders if one of its own properties changes. This
increases the performance for big forms as lot.
You can still activate the old behaviour with the new
"updateOnEveryFormChange" option. This option is available in the
"createField" function.

BREAKING CHANGE: A Field now only updates if one of the own properties
changes. If you had Field that rendered based on other values in the
form, you need to activate the "updateOnEveryFormChange" option.