fix: do not throw error on deep updates outside of strict mode
Then the strict mode is disabled, it is possible
to pass values with missing keys. If you wanted to
update these keys you got an error. I removed
the key check, so no error is thrown in this case.
If the strict mode is enabled, this condition is already
checked on read.